expose predicate methods
The predicate methods were in LOCAL scope. However, for the purposes of
self testing this library it makes sense to expose them to the GLOBAL
scope. If I can do better black-box testing of the ASSERT methods I may
be able to bring the predicate methods back into LOCAL scope.